Move ROCm compute to its own module.
This is in preperation for supporting several compute APIs such as
WebGPU, Vulkan, CUDA, etc.

Even before this change, ROCm no longer works on my Arch Linux system
with a Radeon 580X graphics card. Support had been dropped for the 580x
previously but it still seemed to work with some workarounds. Now, it
does not work at all, hence the reason for moving to a more stable
compute API such as WebGPU or Vulkan.